+++ CHANGELOG.md
@@ -1,5 +1,7 @@
+## 1.2.1
+* change: interceptors.response onFulfilled function parameter (data, config, response) => (data, {url, config}, response)
 ## 1.2.0
-add: interceptors.use
-add: interceptors preset abortRequest
+* add: interceptors.use
+* add: interceptors preset abortRequest
 ## 1.1.0
-add: multi interceptors
+* add: multi interceptors